added changes that allow for multiple loss functions per output... commits missed because of corruption
