--- lightwood/lightwood/__about__.py
@@ -1,10 +1,0 @@
-__title__ = 'lightwood'
-__package_name__ = 'mindsdb'
-__version__ = '0.17.1'
-__description__ = "Lightwood's goal is to make it very simple for developers to use the power of artificial neural networks in their projects."
-__email__ = "jorge@mindsdb.com"
-__author__ = 'MindsDB Inc'
-__github__ = 'https://github.com/mindsdb/lightwood'
-__pypi__ = 'https://pypi.org/project/lightwood'
-__license__ = 'MIT'
-__copyright__ = 'Copyright 2019- mindsdb'
