--- lightwood/lightwood/__init__.py
@@ -1,16 +1,0 @@
-import torchvision
-import sys
-
-
-if sys.version_info < (3, 6):
-    sys.exit('Sorry, For Lightwood Python < 3.6 is not supported')
-
-from lightwood.__about__ import __package_name__ as name, __version__
-from lightwood.encoders import BUILTIN_ENCODERS
-from lightwood.mixers import BUILTIN_MIXERS
-from lightwood.api.predictor import Predictor
-import lightwood.model_building
-import lightwood.constants.lightwood as CONST
-
-
-COLUMN_DATA_TYPES = CONST.COLUMN_DATA_TYPES
