--- lightwood/lightwood/encoders/audio/__init__.py
@@ -1,3 +1,0 @@
-from lightwood.encoders.audio.audio_ts import AmplitudeTsEncoder
-
-default = AmplitudeTsEncoder
