--- lightwood/lightwood/encoders/categorical/__init__.py
@@ -1,6 +1,0 @@
-from lightwood.encoders.categorical.onehot import OneHotEncoder
-from lightwood.encoders.categorical.autoencoder import CategoricalAutoEncoder
-
-default = CategoricalAutoEncoder
-onehot = OneHotEncoder
-autoencoder = CategoricalAutoEncoder
