--- lightwood/lightwood/encoders/datetime/__init__.py
@@ -1,4 +1,0 @@
-from lightwood.encoders.datetime.datetime import DatetimeEncoder
-
-
-default = DatetimeEncoder
