--- lightwood/lightwood/encoders/image/__init__.py
@@ -1,2 +1,0 @@
-from lightwood.encoders.image.img_2_vec import Img2VecEncoder
-default = Img2VecEncoder
