--- lightwood/lightwood/encoders/numeric/__init__.py
@@ -1,2 +1,0 @@
-from lightwood.encoders.numeric.numeric import NumericEncoder
-default = NumericEncoder
