--- lightwood/lightwood/mixers/helpers/qunatile_loss.py
@@ -1,23 +1,0 @@
-import torch
-
-
-class QuantileLoss(torch.nn.Module):
-    def __init__(self, quantiles=[0.85,0.95]):
-        super().__init__()
-        self.quantiles = quantiles
-
-    def forward(self, preds, target):
-        assert not target.requires_grad
-        assert preds.size(0) == target.size(0)
-
-        losses = []
-        for i, q in enumerate(self.quantiles):
-            errors = target - preds[:, i+3]
-            losses.append(
-                torch.max(
-                   (q-1) * errors,
-                   q * errors
-            ).unsqueeze(1))
-        loss = torch.mean(
-            torch.sum(torch.cat(losses, dim=1), dim=1))
-        return loss
