--- lightwood/lightwood/model_building/__init__.py
@@ -1,1 +1,0 @@
-from lightwood.model_building.basic_ax_optimizer.basic_ax_optimizer import BasicAxOptimizer
