+++ lightwood/mixers/helpers/transformer.py
@@ -9,6 +9,7 @@ class Transformer:
         self.output_features = output_features
 
         self.feature_len_map = {}
+        self.out_indexes = []
 
     def transform(self, sample):
 
@@ -27,6 +28,12 @@ class Transformer:
             if output_feature not in self.feature_len_map:
                 self.feature_len_map[output_feature] = len(sub_vector)
 
+            if len(self.out_indexes) < len(sample['output_features']):
+                if len(self.out_indexes) == 0:
+                    self.out_indexes.append([0,len(sub_vector)])
+                else:
+                    self.out_indexes.append([self.out_indexes[-1][1], self.out_indexes[-1][1] + len(sub_vector)])
+                
         return torch.FloatTensor(input_vector), torch.FloatTensor(output_vector)
 
     def revert(self, vector, feature_set='output_features'):
